move core features into their own folder
