+++ code-examples/core_features/did/3_did.ts
