+++ code-examples/core_features/did/4_did.ts
