+++ code-examples/core_features/did/6_did.ts
