+++ code-examples/core_features/did/7_did.ts
