+++ code-examples/core_features/did/8_did.ts
