+++ code-examples/core_features/did/index.ts
@@ -2,11 +2,7 @@ import { randomAsHex } from '@polkadot/util-crypto'
 
 import { BlockchainUtils } from '@kiltprotocol/chain-helpers'
 import { init as kiltInit } from '@kiltprotocol/core'
-import {
-  FullDidDetails,
-  DemoKeystore,
-  DefaultResolver,
-} from '@kiltprotocol/did'
+import { FullDidDetails, DemoKeystore, DefaultResolver } from '@kiltprotocol/did'
 import * as Kilt from '@kiltprotocol/sdk-js'
 
 import { main as main1 } from './1_did'
@@ -47,70 +43,35 @@ export async function runAll() {
 
   console.log('main3 - create full DID')
   const randomMini3 = randomAsHex(32)
-  const did3 = await main3(
-    keystore,
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    randomMini3
-  )
+  const did3 = await main3(keystore, faucetAccount, BlockchainUtils.IS_IN_BLOCK, randomMini3)
 
-  console.log(
-    'main4 - create full DID with encryption key and service endpoints'
-  )
+  console.log('main4 - create full DID with encryption key and service endpoints')
   const authSeed4 = randomAsHex(32)
   const encSeed4 = randomAsHex(32)
-  const did4 = await main4(
-    keystore,
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    authSeed4,
-    encSeed4
-  )
+  const did4 = await main4(keystore, faucetAccount, BlockchainUtils.IS_IN_BLOCK, authSeed4, encSeed4)
 
   console.log('main5 - update auth key, remove service endpoint')
   const randomMini5 = randomAsHex(32)
-  await main5(
-    keystore,
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    randomMini5,
-    did4.details as FullDidDetails
-  )
+  await main5(keystore, faucetAccount, BlockchainUtils.IS_IN_BLOCK, randomMini5, did4.details as FullDidDetails)
 
   console.log('main6 - delete DID')
+  let did4_identifier = did4.details?.did
+  if (typeof did4_identifier === 'undefined') {
+    throw 'The did cannot be deleted, we just created it'
+  }
   // Get the updated DID Doc
-  const updatedDidDetails4 = (await (
-    await DefaultResolver.resolveDoc(did4.details.did)
-  )?.details) as FullDidDetails
+  const updatedDidDetails4 = (await (await DefaultResolver.resolveDoc(did4_identifier))?.details) as FullDidDetails
   if (updatedDidDetails4 === undefined) {
     throw 'We just created the did'
   }
-  await main6(
-    keystore,
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    updatedDidDetails4
-  )
+  await main6(keystore, faucetAccount, BlockchainUtils.IS_IN_BLOCK, updatedDidDetails4)
 
   console.log('main7 - claim DID deposit')
-  await main7(
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    did3.details as FullDidDetails
-  )
+  await main7(faucetAccount, BlockchainUtils.IS_IN_BLOCK, did3.details as FullDidDetails)
 
   console.log('main8 - upgrade light DID')
   const randomMini8 = randomAsHex(32)
-  const did8 = await main8(
-    keystore,
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    randomMini8
-  )
+  const did8 = await main8(keystore, faucetAccount, BlockchainUtils.IS_IN_BLOCK, randomMini8)
   console.log('main7 - again, delete DID from main8')
-  await main7(
-    faucetAccount,
-    BlockchainUtils.IS_IN_BLOCK,
-    did8.details as FullDidDetails
-  )
+  await main7(faucetAccount, BlockchainUtils.IS_IN_BLOCK, did8.details as FullDidDetails)
 }
