+++ code-examples/core_features/package.json
@@ -1,0 +1,16 @@
+{
+  "author": "KILT Protocol",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "@kiltprotocol/sdk-js": "^0.25.3",
+    "ts-node": "^10.5.0",
+    "tweetnacl": "^1.0.3",
+    "typescript": "^4.5.5"
+  },
+  "resolutions": {
+    "@polkadot/keyring": "^8.3.2",
+    "@polkadot/util": "^8.3.2",
+    "@polkadot/util-crypto": "^8.3.2"
+  }
+}
