--- code-examples/package.json
@@ -1,16 +1,0 @@
-{
-  "author": "KILT Protocol",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@kiltprotocol/sdk-js": "^0.24.0-17",
-    "ts-node": "^8.10.2",
-    "tweetnacl": "^1.0.3",
-    "typescript": "^3.9.6"
-  },
-  "resolutions": {
-    "@polkadot/keyring": "7.8.2",
-    "@polkadot/util": "7.8.2",
-    "@polkadot/util-crypto": "7.8.2"
-  }
-}
